Refactoring and formatting
