+++ sashimi/hardware/cameras/mock.py
@@ -1,4 +1,4 @@
-from sashimi.hardware.cameras import AbstractCamera
+from sashimi.hardware.cameras.interface import AbstractCamera
 import numpy as np
 import time
 from skimage.measure import block_reduce
@@ -13,7 +13,9 @@ class MockCamera(AbstractCamera):
         self._roi = (0, 0, self._sensor_resolution[0], self._sensor_resolution[1])
         self._frame_rate = 1 / self._exposure_time
         self._binning = 1
-        self.full_mock_image = np.random.randint(0, 65534, size=self._frame_shape, dtype=np.uint16)
+        self.full_mock_image = np.random.randint(
+            0, 65534, size=self._frame_shape, dtype=np.uint16
+        )
         self.current_mock_image = self.full_mock_image
         self.previous_frame_time = None
         self.current_time = time.time_ns()
@@ -45,7 +47,7 @@ class MockCamera(AbstractCamera):
         self._binning = exp_val
         self._frame_shape = (
             self._sensor_resolution[0] // exp_val,
-            self._sensor_resolution[1] // exp_val
+            self._sensor_resolution[1] // exp_val,
         )
         self.prepare_mock_image()
 
@@ -61,11 +63,10 @@ class MockCamera(AbstractCamera):
     def prepare_mock_image(self):
         self.current_mock_image = block_reduce(
             self.full_mock_image[
-            self._roi[0]:self._roi[2],
-            self._roi[1]:self._roi[3]
+                self._roi[0] : self._roi[2], self._roi[1] : self._roi[3]
             ],
             (self._binning, self._binning),
-            func=np.max
+            func=np.max,
         )
 
     def get_frames(self):
